+++ check_runner/src/main/java/com/barkovsky/check_runner/reflection_json/json_mapper/dto/TestDto.java
@@ -1,4 +1,4 @@
-package com.barkovsky.check_runner.json_mapper.dto;
+package com.barkovsky.check_runner.reflection_json.json_mapper.dto;
 
 import java.util.Arrays;
