+++ check_runner/src/main/java/com/barkovsky/check_runner/reflection_json/reflection_task/cache/api/ChoiceCacheFactory.java
@@ -1,8 +1,8 @@
-package com.barkovsky.check_runner.reflection_task.cache.api;
+package com.barkovsky.check_runner.reflection_json.reflection_task.cache.api;
 
-import com.barkovsky.check_runner.reflection_task.cache.LFUCache;
-import com.barkovsky.check_runner.reflection_task.cache.LRUCache;
-import com.barkovsky.check_runner.reflection_task.entity.Student;
+import com.barkovsky.check_runner.reflection_json.reflection_task.cache.LFUCache;
+import com.barkovsky.check_runner.reflection_json.reflection_task.cache.LRUCache;
+import com.barkovsky.check_runner.reflection_json.reflection_task.entity.Student;
 import org.apache.commons.configuration.ConfigurationException;
 import org.apache.commons.configuration.PropertiesConfiguration;
